Introduce VmProcessor
