+++ executor/src/witgen/machines/block_machine.rs
@@ -280,7 +280,6 @@ impl<'a, T: FieldElement> BlockMachine<'a, T> {
             identity_processor,
             &self.identities,
             self.fixed_data,
-            self.row_factory.clone(),
             &self.witness_cols,
         )
         .with_outer_query(OuterQuery::new(left.to_vec(), right));
